+++ tools/tpm2_makecredential.c
@@ -190,8 +190,8 @@ bool tpm2_tool_onstart(tpm2_options **opts) {
 
     const struct option topts[] = {
       {"enc-key"  ,required_argument, NULL, 'e'},
-      {"sec"     ,required_argument, NULL, 's'},
-      {"name"    ,required_argument, NULL, 'n'},
+      {"secret"   ,required_argument, NULL, 's'},
+      {"name"     ,required_argument, NULL, 'n'},
       {"out-file" ,required_argument, NULL, 'o'},
     };