+++ tools/tpm2_nvreadlock.c
@@ -126,10 +126,9 @@ bool tpm2_tool_onstart(tpm2_options **opts) {
         { "index",                required_argument, NULL, 'x' },
         { "hierarchy",       required_argument, NULL, 'a' },
         { "handle-passwd",        required_argument, NULL, 'P' },
-        { "passwdInHex",          no_argument,       NULL, 'X' },
     };
 
-    *opts = tpm2_options_new("x:a:P:X", ARRAY_LEN(topts), topts,
+    *opts = tpm2_options_new("x:a:P:", ARRAY_LEN(topts), topts,
                              on_option, NULL, TPM2_OPTIONS_SHOW_USAGE);
 
     return *opts != NULL;